% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_helpers.R
\name{add_chronogram_axis}
\alias{add_chronogram_axis}
\title{Add chronogram axis}
\usage{
add_chronogram_axis(
  p,
  tree,
  by = 5,
  digits = 0L,
  axis_offset_frac = 0.052,
  tick_height_frac = 0.01,
  label_offset_frac = 0.028,
  title_margin_top = 28,
  bar_size = 12,
  segment_size = 2.5
)
}
\description{
Add chronogram axis
}
\keyword{internal}
