% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_helpers.R
\name{add_clade_labels_by_level}
\alias{add_clade_labels_by_level}
\title{Add clade labels by level}
\usage{
add_clade_labels_by_level(
  p,
  label_tbl,
  fontsize = 3,
  barsize = 0.45,
  offset = 0.03,
  offset_text = 0.004,
  fontface = "plain",
  sort_desc = TRUE,
  angle = 0,
  align = TRUE
)
}
\description{
Add clade labels by level
}
\keyword{internal}
