% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_helpers.R
\name{apply_clade_label_layers}
\alias{apply_clade_label_layers}
\title{Apply clade label layers}
\usage{
apply_clade_label_layers(p, registry, layer_specs)
}
\description{
Apply clade label layers
}
\keyword{internal}
