% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_helpers.R
\name{assert_constraint_columns}
\alias{assert_constraint_columns}
\title{Assert constraint columns}
\usage{
assert_constraint_columns(constraints_tbl)
}
\description{
Assert constraint columns
}
\keyword{internal}
