% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_helpers.R
\name{augment_treedata_with_registry}
\alias{augment_treedata_with_registry}
\title{Augment treedata with registry annotations}
\usage{
augment_treedata_with_registry(treedata_obj, registry)
}
\arguments{
\item{treedata_obj}{A treedata object}

\item{registry}{A registry data frame}
}
\description{
Augment treedata with registry annotations
}
