% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{automate_treePL}
\alias{automate_treePL}
\title{Automate treePL bootstrap pipeline}
\usage{
automate_treePL(
  cfg_file,
  wrapper_sh,
  ml_tree_file,
  bs_trees_file,
  results_dir,
  treePL_out,
  num_bs = NULL
)
}
\arguments{
\item{cfg_file}{Character. Configuration file.}

\item{wrapper_sh}{Character. Wrapper script path.}

\item{ml_tree_file}{Character. Path to the maximum-likelihood tree.}

\item{bs_trees_file}{Character. Bootstraps tree path.}

\item{results_dir}{Character. Results output dir.}

\item{treePL_out}{Character. Output treePL directory.}

\item{num_bs}{Integer or NULL. Number of bootstrap trees to use. If NULL, uses all available.}
}
\description{
Automate treePL bootstrap pipeline
}
