% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_helpers.R
\name{build_annotation_registry}
\alias{build_annotation_registry}
\title{Build annotation registry}
\usage{
build_annotation_registry(
  tree,
  constraints_tbl,
  main_level4_values = NULL,
  supp_level4_min_tips = 10L,
  require_monophyly = TRUE
)
}
\description{
Build annotation registry
}
\keyword{internal}
