% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{calculate_ml_tree}
\alias{calculate_ml_tree}
\title{Execute ML Search in RAxML-NG}
\usage{
calculate_ml_tree(
  raxml_bin_path,
  aln_file,
  part_file,
  constraint_file,
  outgroup = NULL,
  n_init_trees = "rand{25},pars{25}",
  seed = 111,
  n_workers = 1,
  threads = 4,
  output_dir = dirname(aln_file),
  prefix = "cactus_search"
)
}
\arguments{
\item{raxml_bin_path}{Character. System RAxML-NG path.}

\item{aln_file}{Character. PHYLIP matrix path.}

\item{part_file}{Character. Partitions model path.}

\item{constraint_file}{Character. Constraints scaffold tree path.}

\item{outgroup}{Character. Outgroup species name to root the tree. Defaults to NULL.}

\item{n_init_trees}{Character. Initial trees configuration for search. Defaults to "rand{25},pars{25}".}

\item{seed}{Integer. Seed for reproducibility. Defaults to 111.}

\item{n_workers}{Integer. Number of workers. Defaults to 1.}

\item{threads}{Integer. Run threads.}

\item{output_dir}{Character. Directory to write outputs. Defaults to aln_file directory.}

\item{prefix}{Character. Prefix for output files. Defaults to "cactus_search".}
}
\value{
List of output paths.
}
\description{
Execute ML Search in RAxML-NG
}
