% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{calculate_rf_distances}
\alias{calculate_rf_distances}
\title{Compute Robinson-Foulds Distances between Maximum-Likelihood Trees}
\usage{
calculate_rf_distances(
  raxml_bin_path,
  ml_trees_file,
  output_dir = dirname(ml_trees_file),
  prefix = "cactus_RF"
)
}
\arguments{
\item{raxml_bin_path}{Character. System RAxML-NG path.}

\item{ml_trees_file}{Character. Path to .raxml.mlTrees file.}

\item{output_dir}{Character. Directory for outputs. Defaults to ml_trees_file directory.}

\item{prefix}{Character. Output prefix. Defaults to "cactus_RF".}
}
\value{
Output prefix distance path.
}
\description{
Compute Robinson-Foulds Distances between Maximum-Likelihood Trees
}
