% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{calculate_temporal_bootstraps}
\alias{calculate_temporal_bootstraps}
\title{Compute Temporal Best Maximum-Likelihood Tree Constrained Bootstrap Replicates}
\usage{
calculate_temporal_bootstraps(
  raxml_bin_path,
  aln_file,
  part_file,
  best_tree_file,
  bs_trees = 500,
  outgroup = NULL,
  seed = 111,
  threads = 4,
  output_dir = dirname(aln_file),
  prefix = "cactus_temporal"
)
}
\arguments{
\item{raxml_bin_path}{Character. System RAxML-NG path.}

\item{aln_file}{Character. Multiple sequence alignment path.}

\item{part_file}{Character. Partitions model path.}

\item{best_tree_file}{Character. Best tree topology path (used as constraint).}

\item{bs_trees}{Integer. Bootstrap replicates for temporal analysis. Defaults to 500.}

\item{outgroup}{Character. Outgroup taxon. Defaults to NULL.}

\item{seed}{Integer. Seed. Defaults to 111.}

\item{threads}{Integer. Run threads.}

\item{output_dir}{Character. Directory to write outputs.}

\item{prefix}{Character. Output prefix. Defaults to "cactus_temporal".}
}
\value{
Output bootstrap trees file.
}
\description{
Compute Temporal Best Maximum-Likelihood Tree Constrained Bootstrap Replicates
}
