% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{check_bs_convergence}
\alias{check_bs_convergence}
\title{Check Bootstrap Convergence}
\usage{
check_bs_convergence(
  raxml_bin_path,
  bs_trees_file,
  bs_cutoff = 0.03,
  seed = 111,
  threads = 4,
  output_dir = dirname(bs_trees_file),
  prefix = "cactus_bs_convergence"
)
}
\arguments{
\item{raxml_bin_path}{Character. RAxML-NG bin path.}

\item{bs_trees_file}{Character. Path to concatenated bootstrap trees file.}

\item{bs_cutoff}{Numeric. Cutoff for convergence. Defaults to 0.03.}

\item{seed}{Integer. Seed. Defaults to 111.}

\item{threads}{Integer. Threads. Defaults to 4.}

\item{output_dir}{Character. Directory to write logs. Defaults to dirname(bs_trees_file).}

\item{prefix}{Character. Prefix to use. Defaults to "cactus_bs_convergence".}
}
\value{
Path to convergence log.
}
\description{
Check Bootstrap Convergence
}
