% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{clean_taxonomic_names}
\alias{clean_taxonomic_names}
\title{Reconcile and Validate Taxonomic Names}
\usage{
clean_taxonomic_names(
  raw_input_fasta,
  checklist_path,
  output_clean_dir,
  force_process = FALSE
)
}
\arguments{
\item{raw_input_fasta}{Character path. Raw download coordinates folder.}

\item{checklist_path}{Character. Excel Cactaceae accepted spelling table file.}

\item{output_clean_dir}{Character. Folders for cleaned configurations.}

\item{force_process}{Logical. Skip cache check.}
}
\description{
Reconcile and Validate Taxonomic Names
}
