% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{collect_bootstraps}
\alias{collect_bootstraps}
\title{Collect Bootstrap Trees}
\usage{
collect_bootstraps(
  bs_dir,
  output_dir = bs_dir,
  prefix = "cactus_ALL_bootstraps"
)
}
\arguments{
\item{bs_dir}{Character. Directory where bootstraps chunks are stored.}

\item{output_dir}{Character. Directory to write concatenated bootstraps. Defaults to bs_dir.}

\item{prefix}{Character. Prefix to use. Defaults to "cactus_ALL_bootstraps".}
}
\value{
Path to concatenated trees file.
}
\description{
Collect Bootstrap Trees
}
