% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_helpers.R
\name{compute_group_nodes}
\alias{compute_group_nodes}
\title{Compute group nodes}
\usage{
compute_group_nodes(
  tree,
  tip_tbl,
  group_col,
  group_values = NULL,
  min_tips = 2L
)
}
\description{
Compute group nodes
}
\keyword{internal}
