% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{generate_bootstrap_script}
\alias{generate_bootstrap_script}
\title{Generate Chunk Bootstrap Shell Script}
\usage{
generate_bootstrap_script(
  alignment_file,
  partition_file,
  constraint_file,
  outgroup = NULL,
  bs_per_rep = 500,
  max_reps = 4,
  base_seed = 111,
  seed_step = 1000,
  threads = 120,
  workers = 20,
  output_dir = getwd(),
  script_name = "run_bs_chunks.sh",
  cluster_job_name = "C-raxml.bs",
  cluster_mem = "32G",
  cluster_time = "7-00:00",
  cluster_partition = "general",
  cluster_queue = "public",
  load_module = "raxml-ng-1.1.0-gcc-11.2.0",
  raxml_exec = "raxml-ng-mpi"
)
}
\arguments{
\item{alignment_file}{Character. Alignment file path.}

\item{partition_file}{Character. Partition model file path.}

\item{constraint_file}{Character. Constraint tree file path.}

\item{outgroup}{Character. Outgroup taxon. Defaults to NULL.}

\item{bs_per_rep}{Integer. Bootstraps per replicate. Defaults to 500.}

\item{max_reps}{Integer. Number of chunks. Defaults to 4.}

\item{base_seed}{Integer. Base seed. Defaults to 111.}

\item{seed_step}{Integer. Step in seed increments per chunk. Defaults to 1000.}

\item{threads}{Integer. Threads per task. Defaults to 120.}

\item{workers}{Integer. Workers per task. Defaults to 20.}

\item{output_dir}{Character. Where to output scripts and chunks.}

\item{script_name}{Character. The name of the bash file.}

\item{cluster_job_name}{Character. Job name for SLURM. Defaults to "C-raxml-bs".}

\item{cluster_mem}{Character. Memory for SLURM. Defaults to "32G".}

\item{cluster_time}{Character. Time limit for SLURM. Defaults to "7-00:00".}

\item{cluster_partition}{Character. Partition for SLURM. Defaults to "general".}

\item{cluster_queue}{Character. Queue for SLURM. Defaults to "public".}

\item{load_module}{Character. Module to load in SLURM script. Defaults to "raxml-ng-1.1.0-gcc-11.2.0".}

\item{raxml_exec}{Character. RAxML executable to use. Defaults to "raxml-ng-mpi".}
}
\value{
The path to the generated script.
}
\description{
Generate Chunk Bootstrap Shell Script
}
