% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_helpers.R
\name{get_annotation_nodes}
\alias{get_annotation_nodes}
\title{Get annotation nodes}
\usage{
get_annotation_nodes(
  tree,
  constraints_tbl,
  group_col,
  group_values = NULL,
  min_tips = 2L,
  require_monophyly = TRUE
)
}
\description{
Get annotation nodes
}
\keyword{internal}
