% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn.R
\name{get_iucn_data}
\alias{get_iucn_data}
\title{Extract IUCN Data from rredlist}
\usage{
get_iucn_data(sp_name)
}
\arguments{
\item{sp_name}{The scientific name of the species}
}
\value{
A tibble with extracted IUCN metadata
}
\description{
Extract IUCN Data from rredlist
}
