% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{integrate_and_clean_markers}
\alias{integrate_and_clean_markers}
\title{Final Marker Integration and Taxonomic Cleaning}
\usage{
integrate_and_clean_markers(
  ingroup_dir,
  outgroup_dir,
  output_dir,
  accepted_list_file,
  metadata_in_file,
  metadata_out_file
)
}
\arguments{
\item{ingroup_dir}{Character.}

\item{outgroup_dir}{Character.}

\item{output_dir}{Character.}

\item{accepted_list_file}{Character.}

\item{metadata_in_file}{Character.}

\item{metadata_out_file}{Character.}
}
\description{
Final Marker Integration and Taxonomic Cleaning
}
