% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{map_branch_supports}
\alias{map_branch_supports}
\title{Calculate and Burn Branch Supports}
\usage{
map_branch_supports(
  raxml_bin,
  best_tree,
  bootstraps_file,
  metric = "tbe",
  threads = 4,
  output_dir = dirname(best_tree),
  prefix = "cactus_support"
)
}
\arguments{
\item{raxml_bin}{Character. System RAxML-NG path.}

\item{best_tree}{Character. Best tree topology path.}

\item{bootstraps_file}{Character. Replicates tree file path.}

\item{metric}{Character. Support type ('tbe' or 'fbp').}

\item{threads}{Integer. Calculations threads count.}

\item{output_dir}{Character. Directory to write the support file. Defaults to dirname(best_tree).}

\item{prefix}{Character. Output prefix. Defaults to "cactus_support".}
}
\value{
Character path to the written support tree.
}
\description{
Calculate and Burn Branch Supports
}
