% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_helpers.R
\name{prepare_tip_annotation}
\alias{prepare_tip_annotation}
\title{Prepare tip annotation}
\usage{
prepare_tip_annotation(tree, constraints_tbl)
}
\description{
Prepare tip annotation
}
\keyword{internal}
