% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{preprocess_partitions}
\alias{preprocess_partitions}
\title{Preprocess Partitions and Run Alignment Check}
\usage{
preprocess_partitions(
  phy_matrix,
  part_file,
  raxml_path,
  output_dir = dirname(phy_matrix),
  force_check = FALSE
)
}
\arguments{
\item{phy_matrix}{Character. Input PHYLIP file path.}

\item{part_file}{Character. Input partition mapping file.}

\item{raxml_path}{Character. Location of the RAxML-NG executable.}

\item{output_dir}{Character. Folders for validated partitions outputs. Defaults to dirname(phy_matrix).}

\item{force_check}{Logical. Bypass run checks if validation exists.}
}
\description{
Preprocess Partitions and Run Alignment Check
}
