% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{rescale_tree}
\alias{rescale_tree}
\title{Rescale all branches of a tree}
\usage{
rescale_tree(tree, factor = 100)
}
\arguments{
\item{tree}{phylo.}

\item{factor}{numeric.}
}
\value{
phylo
}
\description{
Rescale all branches of a tree
}
