% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenation.R
\name{run_concatenation_pipeline}
\alias{run_concatenation_pipeline}
\title{Run Concatenation Pipeline}
\usage{
run_concatenation_pipeline(input_dir, output_dir)
}
\arguments{
\item{input_dir}{Character.}

\item{output_dir}{Character.}
}
\description{
Run Concatenation Pipeline
}
