% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{run_joint_realignment}
\alias{run_joint_realignment}
\title{Run Final Joint Realignment Pipeline}
\usage{
run_joint_realignment(input_dir, output_fasta_dir, output_aln_dir)
}
\arguments{
\item{input_dir}{Character.}

\item{output_fasta_dir}{Character.}

\item{output_aln_dir}{Character.}
}
\description{
Run Final Joint Realignment Pipeline
}
