% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{run_local_bootstraps}
\alias{run_local_bootstraps}
\title{Run Bootstrap Search Locally}
\usage{
run_local_bootstraps(
  raxml_bin_path,
  aln_file,
  part_file,
  constraint_file,
  bs_trees = 500,
  outgroup = NULL,
  seed = 111,
  threads = 8,
  workers = 1,
  output_dir = dirname(aln_file),
  prefix = "cactus_bs"
)
}
\arguments{
\item{raxml_bin_path}{Character. System RAxML-NG path.}

\item{aln_file}{Character. Multiple sequence alignment path.}

\item{part_file}{Character. Partitions model path.}

\item{constraint_file}{Character. Constraint tree path.}

\item{bs_trees}{Integer. Total number of bootstrap trees to generate. Defaults to 500.}

\item{outgroup}{Character. Outgroup taxon. Defaults to NULL.}

\item{seed}{Integer. Seed. Defaults to 111.}

\item{threads}{Integer. Run threads. Defaults to 4.}

\item{workers}{Integer. Number of workers. Defaults to 1.}

\item{output_dir}{Character. Directory to write outputs.}

\item{prefix}{Character. Output prefix. Defaults to "cactus_bs".}
}
\value{
Output bootstrap trees file.
}
\description{
Run Bootstrap Search Locally
}
