% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{run_mafft}
\alias{run_mafft}
\title{Run MAFFT Alignment}
\usage{
run_mafft(
  input_fasta,
  output_fasta,
  mafft_exec = "mafft",
  mafft_opts = "--auto"
)
}
\arguments{
\item{input_fasta}{Character. Input FASTA file path.}

\item{output_fasta}{Character. Output aligned FASTA path.}

\item{mafft_exec}{Character. System MAFFT binary location.}

\item{mafft_opts}{Character. Parameters passed to MAFFT.}
}
\value{
Numeric. Exit status code.
}
\description{
Run MAFFT Alignment
}
