% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{run_marker_screening}
\alias{run_marker_screening}
\title{Run Ingroup Marker Screening Pipeline}
\usage{
run_marker_screening(
  fasta_folder,
  out_base,
  min_cols_to_evaluate = 50L,
  min_aln_len_to_retain = 200L,
  min_nseq_to_retain = 100L,
  max_marker_missing = 0.7,
  saturation_flag_cutoff = 0.3,
  saturation_keep_cutoff = 0.5,
  iqr_multiplier = 1.5
)
}
\arguments{
\item{fasta_folder}{Character. Directory containing aligned FASTAs.}

\item{out_base}{Character. Base directory for output files.}

\item{min_cols_to_evaluate}{Integer. Minimum columns to evaluate.}

\item{min_aln_len_to_retain}{Integer. Minimum alignment length to retain.}

\item{min_nseq_to_retain}{Integer. Minimum sequences to retain.}

\item{max_marker_missing}{Numeric. Maximum missing fraction for a marker.}

\item{saturation_flag_cutoff}{Numeric.}

\item{saturation_keep_cutoff}{Numeric. Retention threshold for the saturation slope.}

\item{iqr_multiplier}{Numeric. Scale factor for IQR outlier bounds detection.}
}
\value{
Summarized table.
}
\description{
Run Ingroup Marker Screening Pipeline
}
