% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{run_modeltest_ng}
\alias{run_modeltest_ng}
\title{Run ModelTest-NG Partition Evaluations}
\usage{
run_modeltest_ng(
  modeltest_exec_path,
  aln_file,
  part_file,
  prefix = "MODELTEST_cactus_phylo",
  threads = 4
)
}
\arguments{
\item{modeltest_exec_path}{Character. Path to ModelTest-NG.}

\item{aln_file}{Character. Checked PHYLIP alignment path.}

\item{part_file}{Character. Cleaned partitions path.}

\item{prefix}{Character. File output prefix.}

\item{threads}{Integer. Processing threads count.}
}
\value{
Character path to the resulting partition map file.
}
\description{
Run ModelTest-NG Partition Evaluations
}
