% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{run_treePL}
\alias{run_treePL}
\title{Run treePL via Wrapper}
\usage{
run_treePL(cfg, treefile, label, wrapper_sh)
}
\arguments{
\item{cfg}{Character. Configuration file.}

\item{treefile}{Character. Tree file.}

\item{label}{Character. Label for the run.}

\item{wrapper_sh}{Character. Path to the treepl wrapper shell script.}
}
\description{
Cites the reference treePL wrapper: https://github.com/tongjial/treepl_wrapper
}
