% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{run_treePL_direct}
\alias{run_treePL_direct}
\title{Run treePL directly}
\usage{
run_treePL_direct(cfg_file, label, cwd = NULL)
}
\arguments{
\item{cfg_file}{Character. Configuration file.}

\item{label}{Character. Label.}

\item{cwd}{Character. Working directory.}
}
\description{
Run treePL directly
}
