% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn.R
\name{scale_fill_iucn}
\alias{scale_fill_iucn}
\title{IUCN Category Color Scale for ggplot2}
\usage{
scale_fill_iucn(...)
}
\arguments{
\item{...}{arguments passed to scale_fill_manual}
}
\description{
IUCN Category Color Scale for ggplot2
}
