% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_phylogenies}
\alias{validate_phylogenies}
\title{Comparative Tree space validation pipeline}
\usage{
validate_phylogenies(
  trees_mapping_list,
  checklist_csv,
  constraints_map,
  out_dir
)
}
\arguments{
\item{trees_mapping_list}{Named list of trees.}

\item{checklist_csv}{Character. Spellings check list CSV.}

\item{constraints_map}{Character. Taxonomy mapping table.}

\item{out_dir}{Character. Destination folder.}
}
\value{
Generates tree distance results and tree-space coordinates.
}
\description{
Comparative Tree space validation pipeline
}
